backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI & ML
anthropic>=0.39.0
sentence-transformers==2.3.1
chromadb==0.4.22
langchain==0.1.6
langchain-community==0.0.19

# Document Processing
PyMuPDF==1.23.21
pdfplumber==0.10.3
camelot-py[cv]==0.11.0
pillow==10.2.0
pytesseract==0.3.10
openpyxl==3.1.2
pandas==2.2.0
python-docx==1.1.0

# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
feedparser==6.0.11

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.6
boto3==1.34.34  # S3 uploads (Sprint 12)

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1


